Move 'artifacts' classes to artifacts module
