+++ artifacts/src/main/groovy/com/novoda/gradle/release/artifacts/AndroidLibrary.groovy
@@ -1,4 +1,4 @@
-package com.novoda.gradle.release
+package com.novoda.gradle.release.artifacts
 
 import org.gradle.api.DomainObjectSet
 import org.gradle.api.Project
